Session 04 Iterations
In this season we created 5 classes which each of one have their methods. 
The first Class is the Reverse which takes a string and reverse it using the method ReverseString.
The Second class is the calculator which has one method to calculate the sum from 1 to n. The other method calculates the product from 1 to n. The Third Class is the Prime which finds the prime numbers between 1 and n. The Fourth class is the ArraysCalculator which has a method to calculate arrays multiplication and one method to print one array. Last but not least we have the Sort class which implements  Select Sort algorithm