Session-07 Ui University 1.3
New :
Slove the bug with listView by convert them to chexkBox List, it makes the Ui unglier but it works.
EidtForm works for Profesoor/student/ course and can fill all the textFields

Next update:
Add action to  buttons  New, Delete and Save 
Add action to MenuButtons for the Grades and schedule.